Add House and Character model into QuoteModel. Minimize amount of files.
