--- app/src/main/java/com/jojjnator/kalkylator/model/Character.kt
@@ -1,10 +1,0 @@
-package com.jojjnator.kalkylator.model
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Character(
-    val house: House,
-    val name: String,
-    val slug: String
-)
