--- app/src/main/java/com/jojjnator/kalkylator/model/House.kt
@@ -1,10 +1,0 @@
-package com.jojjnator.kalkylator.model
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class House(
-
-    val name: String,
-    val slug: String
-)
