+++ app/src/main/java/com/jojjnator/kalkylator/model/QuoteModel.kt
@@ -4,7 +4,19 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class QuoteModel(
-
     val character: Character,
     val sentence: String
+)
+
+@Serializable
+data class House(
+    val name: String,
+    val slug: String
+)
+
+@Serializable
+data class Character(
+    val house: House,
+    val name: String,
+    val slug: String
 )
