+++ app/src/main/java/com/jojjnator/kalkylator/network/KtorClient.kt
@@ -15,7 +15,4 @@ class KtorClient {
             })
         }
     }
-
-
-
 }
